Prachet dev (#21)
* Further changes on login and account page

* Update Navbar.component.jsx

* Committing working changes

* Fixing cookies bug and adding some more code related to booking page

* including CSS changes that were overwritten

* linted and fixed various syntax

I think we're using different configurations of ESLint. Let's chat about it and get on the same page

* fixed build issues from using anchor link to next/link

* Updating NPM

* Test commit

* Update route-choices.jsx

* Tailwind install

* Changes for Autofill on the checkout page

* Further changes

Added protected routes, so that user can only access certain routes if they are logged in.

Added cities in checkout order summary.

* Update MapSection.component.jsx

Fixing error

* Update checkout.jsx

Adding changes for taxes being dynamic, payment form, checking the correct card info, autofill.

* Update checkout.jsx

changing the dynamic total cost to the submit button

* Adding the getTrips service

* Changes

* Made changes on account.jsx

Added logic for fetching the name & all trips for the logged in user. Also added logic to show a message if no trips have been booked for that user.

* Changes to add a trip on successful payment

Adding logic to add a trip to db on a successful payment. Routing to account page on successful addition of a trip

* Update RouteOptionCard.component.jsx

* Fixing build bugs

* Update withAuth.jsx

* added key from pods array to Reservation component

* Fixing build errors

* Update account.jsx

Co-authored-by: Brian Prost <brian.s.prost@gmail.com>